control: clean up
